Add backup to more files
ToDo: Create backup and link function to avoid repetition

+++ setup-links.sh
@@ -12,11 +12,11 @@ _sed() {
 #_sed "/^additional_files=/s#\"[^\"]*\"#\"$PWD\"#g" $PWD/bash/bashrc
 
 # Bash files
-if [ -f  ~/.bashrc ]; then
-    mv ~/.bashrc ~/.bashrc_backup
-fi;
+[[ -f  ~/.bashrc && ! -h ~/.bashrc ]] && mv ~/.bashrc ~/.bashrc_backup
 ln -sv $PWD/bash/bashrc ~/.bashrc
+[[ -f ~/.bash_profile && ! -h ~/.bash_profile ]] && mv ~/.bash_profile ~/.bash_profile_backup
 ln -sv $PWD/bash/bash_profile ~/.bash_profile
+[[ -f ~/.bash_logout && ! -h ~/.bash_logout ]] && mv ~/.bash_logout ~/.bash_logout_backup
 ln -sv $PWD/bash/bash_logout ~/.bash_logout
 
 if [ "$TERM_PROGRAM" == "Apple_Terminal" ]; then
@@ -24,6 +24,7 @@ if [ "$TERM_PROGRAM" == "Apple_Terminal" ]; then
 fi
 
 # screen files
+[[ -f ~/.tmux.conf && ! -h ~/.tmux.conf ]] && mv ~/.tmux.conf ~/.tmux.conf_backup
 ln -sv $PWD/tmux/tmux.conf ~/.tmux.conf
 
 # inputrc. For programs that use readline library
@@ -34,6 +35,7 @@ ln -sv $PWD/bash/editrc ~/.editrc
 
 
 # git files
+[[ -f ~/.gitconfig && ! -h ~/.gitconfig ]] && mv ~/.gitconfig ~/.gitconfig_backup
 ln -sv $PWD/git/gitconfig ~/.gitconfig       # You might have to adjust some paths in this file
 
 # vim files
